refactor(composer): Composer.tsx 拆入 Composer/ 子目录
- 4127→2066 行：types 283/utils 78/composerMemo 71/ContextStack 376 + hooks×7
- memo 比较器字段顺序逐字；usage useMemo 壳与 deps 逐字；target 变更回调留主文件
- ComposerRewindDialogRequest 等 4 公共类型主文件 re-export（useLayoutNodes 跨包契约不变）
- baseline 重生成 + 新文件登记豁免
验证：tsc 零 error；13 测试文件 117 passed，9 失败与基线逐一同名（既有）；gate 绿
（子代理 ComposerSplit 执行，主代理复核验证）
