+++ docs/architecture/large-file-baseline.json
@@ -1,5 +1,5 @@
 {
-  "generatedAt": "2026-08-30T19:15:50.955Z",
+  "generatedAt": "2026-08-30T19:24:47.279Z",
   "scope": "fail",
   "policyVersion": "2026-08-30.policy-v5",
   "entries": [
